# docs/openapi.yaml
openapi: 3.1.0
info:
  title: OmniRoute API
  version: 3.8.50
  description: |
    OmniRoute is a local-first AI API proxy router. It provides an OpenAI-compatible
    endpoint that routes requests to multiple AI providers with load balancing,
    failover, and usage tracking.

    ## Base URLs
    - **Local**: `http://localhost:20128`

    ## Authentication
    All proxy endpoints require a Bearer token (API key managed via the dashboard).
    Management endpoints are protected when `requireLogin` is enabled.
  contact:
    name: OmniRoute
  license:
    name: MIT
servers:
  - url: http://localhost:20128
    description: Local development
...
paths:
...
  /api/settings/cache-config:
    get:
      tags: [Settings]
      summary: Get platform-wide cache configuration
      description: Requires management authentication for the platform-admin tenant.
      security:
        - ManagementSessionAuth: []
      responses:
        "200":
          description: Current cache configuration
        "401":
          description: Unauthorized
        "403":
          description: Platform administrator required
    put:
      tags: [Settings]
      summary: Update platform-wide cache configuration
      description: Requires management authentication for the platform-admin tenant.
      security:
        - ManagementSessionAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                semanticCacheEnabled: { type: boolean }
                semanticCacheMaxSize: { type: number, exclusiveMinimum: 0 }
                semanticCacheTTL: { type: number, exclusiveMinimum: 0 }
                promptCacheEnabled: { type: boolean }
                promptCacheStrategy: { type: string, enum: [auto, system-only, manual] }
                alwaysPreserveClientCache: { type: string, enum: [auto, always, never] }
                idempotencyWindowMs: { type: number, exclusiveMinimum: 0 }
                modelCatalogCacheTtlMs: { type: number, exclusiveMinimum: 0 }
      responses:
        "200":
          description: Cache configuration updated
        "400":
          description: Invalid cache configuration
        "401":
          description: Unauthorized
        "403":
          description: Platform administrator required
...
  /api/rate-limits:
    get:
      tags: [System]
      summary: Get per-account rate limit status
      description: Platform-admin only; returns connection status, model lockouts, and shared cache statistics.
      security:
        - ManagementSessionAuth: []
      responses:
        "200":
          description: Rate limit status by account
        "401": { description: Unauthorized }
        "403": { description: Platform administrator required }
    post:
      tags: [System]
      summary: Toggle per-connection rate-limit protection
      description: Platform-admin only.
      security:
        - ManagementSessionAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [connectionId, enabled]
              properties:
                connectionId: { type: string }
                enabled: { type: boolean }
      responses:
        "200": { description: Rate-limit protection updated }
        "400": { description: Invalid request }
        "401": { description: Unauthorized }
        "403": { description: Platform administrator required }
...
  /api/cache:
    get:
      tags: [System]
      summary: Get cache statistics
      description: Platform-admin only; returns shared semantic-cache and idempotency statistics.
      security:
        - ManagementSessionAuth: []
      responses:
        "200":
          description: Semantic cache and idempotency stats
        "401":
          description: Unauthorized
        "403":
          description: Platform administrator required
    delete:
      tags: [System]
      summary: Clear all caches
      description: Platform-admin only; clears shared cache state.
      security:
        - ManagementSessionAuth: []
      responses:
        "200":
          description: Caches cleared
        "401":
          description: Unauthorized
        "403":
          description: Platform administrator required
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      description: >
        Two bearer families are accepted. Inference API keys (typically `sk-…`)
        authorize `/v1/*`. Management routes also accept `oma_live_…` Access Tokens
        (Settings → Access Tokens / `omniroute connect`) and API keys whose metadata
        includes `manage` or `admin` scope. See docs/guides/MANAGEMENT-AUTH.md.
        Bearer credentials are accepted on management routes that use this scheme;
        they are not rejected solely for being Bearer.
    ManagementSessionAuth:
      type: apiKey
      in: cookie
      name: auth_token
      description: Dashboard management session cookie (auth_token) for protected management routes. Distinct from Bearer Access Tokens and API keys. See docs/guides/MANAGEMENT-AUTH.md.
...